[math][fit] Change FitFCN/SetFCN to pass a flag for the type of fit
In the FitFCN/SetFCN functions use a flag for the fit type instead of passing
a boolean for the chi2 fit.
The flag is by default 0 (undefined type), but can be 1 (for chi2 fit), so compatible as passing a boolean as before, 2 for unbinned likelihood and 3 for binned likelihood (FCN is defined to be the Baker-Cousins log-likelihood ratio)

By passing the correct flag the function FitResult::Chi2() returns the corresponding Chi2() and the Baker-Cousins chi2 equivalent  in case a binned likelihood fits.