[RF] Consistent treatment of empty datasets with new CPU backend
As we discovered in a CMSSW ROOT `master` sync PR, the new RooFit CPU
backend treats empty datasets differently from the legacy NLL evaluation
backend:
https://github.com/cms-sw/cmsdist/pull/9025

This commit is fixing this, in particular removing the assumption that
datasets used for fits with the new CPU backend are never empty.

A unit test that validates the behavior for empty data objects is also
added.