Fix writing long strings.
In TBufferFile::WriteFastArrayString, we had

```
   if (n < 255) {
      *this << (UChar_t)n;
   } else {
      *this << (UChar_t)255;
      *this << n;
   }
```

A recent commit changed the type of the n parameter from Int_t to Long64_t.
This is effectively an incompatible change in the on-disk format, but only for strings
which are at least 255 characters long.

Further, ReadFastArrayString is still reading an Int_t, so this version of ROOT cannot
read files that it writes.

Resolve by changing WriteFastArrayString to explicitly write an Int_t.

Also move the bounds check on the length to before writing anything into the buffer.

Fixes a failure seen in the unit tests of the ATLAS EventInfo package
in the dev3LCG build.