[RF] Don't consider implicit MT in RooBatchCompute
In the original implementation of the RooBatchCompute library, the
evaluation was done multi-threaded in implicit multi-threading was
enabled in ROOT with `ROOT::EnableImplicitMT()`.

This commit removes this feature now, because it was never really
validated, overlapped with other parallelization capabilities of RooFit,
and keeping it will have many unforeseen consequences on users now that
the new CPU evaluation backend is the default in RooFit.

See also:
https://github.com/cms-sw/cmsdist/pull/9047