[core] Silence 6.34 deprecation warnings for now
We are currently discussing what the next ROOT release version number
will be, and if we want to remove the deprecated methods now. To avoid
the annoying warnings, change the check to 6.33.02; *if* we release a
development release later this year, it would likely be called 6.33.01
or 6.33.02. Afterwards master would be bumped to 6.33.03 and we would
again be reminded to remove the methods.